refactored click paint event handling and reorganized utility imports
